services/upload-service/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
azure-storage-blob
azure-eventhub